Del opp TestUtils.kt (1035 linjer) i fire fokuserte filer
TestUtils.kt var ein monolitt med alt mogleg – fake-repositoriar,
motor-fabrikkar, test-klientar og testdata-hjelpefunksjonar.

Del opp langs ansvarslinjer:
- testutils/fakes/FakeRepositories.kt    — alle Fake*-klasser
- testutils/motor/MotorTestHelpers.kt    — motor-fabrikkar og MockJobbAppender
- testutils/builders/TestDataBuilders.kt — opprettTest*, forberedDatabase osv.
- testutils/client/TestClientHelpers.kt  — testKlient*, ventPåSvar*, schemaRegistry

Alle klassane ligg framleis i pakken no.nav.aap.statistikk.testutils, så ingen
importar i eksisterande testar treng endrast.
