--- app/src/test/kotlin/no/nav/aap/statistikk/testutils/TestUtils.kt
@@ -1,1036 +1,0 @@
-package no.nav.aap.statistikk.testutils
-
-import com.google.cloud.NoCredentials
-import com.google.cloud.bigquery.BigQuery
-import com.google.cloud.bigquery.BigQueryOptions
-import com.google.cloud.bigquery.DatasetInfo
-import io.ktor.server.engine.*
-import io.ktor.server.netty.*
-import io.mockk.Runs
-import io.mockk.every
-import io.mockk.just
-import io.mockk.mockk
-import kotlinx.coroutines.runBlocking
-import no.nav.aap.behandlingsflyt.kontrakt.behandling.BehandlingReferanse
-import no.nav.aap.behandlingsflyt.kontrakt.behandling.Status
-import no.nav.aap.behandlingsflyt.kontrakt.behandling.ÅrsakTilOpprettelse
-import no.nav.aap.behandlingsflyt.kontrakt.hendelse.AvklaringsbehovHendelseDto
-import no.nav.aap.behandlingsflyt.kontrakt.hendelse.TilbakekrevingsbehandlingOppdatertHendelse
-import no.nav.aap.behandlingsflyt.kontrakt.statistikk.StoppetBehandling
-import no.nav.aap.komponenter.dbconnect.DBConnection
-import no.nav.aap.komponenter.dbconnect.transaction
-import no.nav.aap.komponenter.gateway.Factory
-import no.nav.aap.komponenter.httpklient.httpclient.ClientConfig
-import no.nav.aap.komponenter.httpklient.httpclient.RestClient
-import no.nav.aap.komponenter.httpklient.httpclient.error.DefaultResponseHandler
-import no.nav.aap.komponenter.httpklient.httpclient.post
-import no.nav.aap.komponenter.httpklient.httpclient.request.PostRequest
-import no.nav.aap.komponenter.httpklient.httpclient.tokenprovider.azurecc.AzureConfig
-import no.nav.aap.komponenter.httpklient.httpclient.tokenprovider.azurecc.ClientCredentialsTokenProvider
-import no.nav.aap.komponenter.repository.RepositoryProvider
-import no.nav.aap.komponenter.type.Periode
-import no.nav.aap.motor.JobbInput
-import no.nav.aap.motor.Motor
-import no.nav.aap.motor.retry.DriftJobbRepositoryExposed
-import no.nav.aap.motor.testutil.ManuellMotorImpl
-import no.nav.aap.motor.testutil.TestJobbRepository
-import no.nav.aap.oppgave.statistikk.OppgaveHendelse
-import no.nav.aap.postmottak.kontrakt.hendelse.DokumentflytStoppetHendelse
-import no.nav.aap.statistikk.AppConfig
-import no.nav.aap.statistikk.avsluttetbehandling.*
-import no.nav.aap.statistikk.behandling.*
-import no.nav.aap.statistikk.beregningsgrunnlag.repository.IBeregningsgrunnlagRepository
-import no.nav.aap.statistikk.bigquery.*
-import no.nav.aap.statistikk.db.DbConfig
-import no.nav.aap.statistikk.db.TransactionExecutor
-import no.nav.aap.statistikk.defaultGatewayProvider
-import no.nav.aap.statistikk.postgresRepositoryRegistry
-import no.nav.aap.statistikk.hendelser.BehandlingService
-import no.nav.aap.statistikk.integrasjoner.pdl.Adressebeskyttelse
-import no.nav.aap.statistikk.integrasjoner.pdl.Gradering
-import no.nav.aap.statistikk.integrasjoner.pdl.PdlGateway
-import no.nav.aap.statistikk.jobber.LagreAvklaringsbehovHendelseJobb
-import no.nav.aap.statistikk.jobber.LagreStoppetHendelseJobb
-import no.nav.aap.statistikk.jobber.appender.JobbAppender
-import no.nav.aap.statistikk.jobber.appender.MotorJobbAppender
-import no.nav.aap.statistikk.meldekort.FritaksvurderingRepository
-import no.nav.aap.statistikk.meldekort.Fritakvurdering
-import no.nav.aap.statistikk.meldekort.IMeldekortRepository
-import no.nav.aap.statistikk.meldekort.Meldekort
-import no.nav.aap.statistikk.module
-import no.nav.aap.statistikk.motor
-import no.nav.aap.statistikk.oppgave.EnhetReservasjonOgTidspunkt
-import no.nav.aap.statistikk.oppgave.LagreOppgaveHendelseJobb
-import no.nav.aap.statistikk.oppgave.LagreOppgaveJobb
-import no.nav.aap.statistikk.oppgave.OppgaveHendelseRepository
-import no.nav.aap.statistikk.oppgave.OppgaveRepositoryImpl
-import no.nav.aap.statistikk.person.IPersonRepository
-import no.nav.aap.statistikk.person.Person
-import no.nav.aap.statistikk.person.PersonRepository
-import no.nav.aap.statistikk.person.PersonService
-import no.nav.aap.statistikk.postmottak.LagrePostmottakHendelseJobb
-import no.nav.aap.statistikk.tilbakekreving.LagreTilbakekrevingHendelseJobb
-import no.nav.aap.statistikk.sak.*
-import no.nav.aap.statistikk.saksstatistikk.*
-import no.nav.aap.statistikk.skjerming.SkjermingService
-import no.nav.aap.statistikk.startUp
-import no.nav.aap.statistikk.tilkjentytelse.TilkjentYtelse
-import no.nav.aap.statistikk.tilkjentytelse.repository.ITilkjentYtelseRepository
-import no.nav.aap.statistikk.tilkjentytelse.repository.TilkjentYtelseEntity
-import no.nav.aap.statistikk.vilkårsresultat.repository.IVilkårsresultatRepository
-import no.nav.aap.statistikk.vilkårsresultat.repository.VilkårsResultatEntity
-import org.slf4j.LoggerFactory
-import org.testcontainers.containers.wait.strategy.HostPortWaitStrategy
-import org.testcontainers.gcloud.BigQueryEmulatorContainer
-import org.testcontainers.postgresql.PostgreSQLContainer
-import org.testcontainers.utility.MountableFile
-import java.io.InputStream
-import java.net.URI
-import java.nio.file.Path
-import java.time.Clock
-import java.time.Duration
-import java.time.LocalDateTime
-import java.time.temporal.ChronoUnit
-import java.util.*
-import javax.sql.DataSource
-import kotlin.system.measureTimeMillis
-
-private val logger = LoggerFactory.getLogger("TestUtils")
-
-/**
- * @param azureConfig Send inn egen her om det skal gjøres autentiserte kall.
- */
-fun <E> testKlient(
-    transactionExecutor: TransactionExecutor,
-    motor: Motor,
-    azureConfig: AzureConfig = AzureConfig(
-        clientId = "tilgang",
-        jwksUri = "http://localhost:8081/jwks",
-        issuer = "tilgang",
-        tokenEndpoint = URI.create("http://localhost:8081/jwks"),
-        clientSecret = "xxx",
-    ),
-    lagreStoppetHendelseJobb: LagreStoppetHendelseJobb,
-    jobbAppender: JobbAppender,
-    test: TestClient.() -> E,
-): E? {
-    val res: E?
-
-    System.setProperty("azure.openid.config.token.endpoint", azureConfig.tokenEndpoint.toString())
-    System.setProperty("azure.app.client.id", azureConfig.clientId)
-    System.setProperty("azure.app.client.secret", azureConfig.clientSecret)
-    System.setProperty("azure.openid.config.jwks.uri", azureConfig.jwksUri)
-    System.setProperty("azure.openid.config.issuer", azureConfig.issuer)
-    val randomUUID = UUID.randomUUID()
-    System.setProperty("integrasjon.postmottak.azp", randomUUID.toString())
-    System.setProperty("integrasjon.oppgave.azp", randomUUID.toString())
-    System.setProperty("integrasjon.behandlingsflyt.azp", randomUUID.toString())
-
-    System.setProperty("NAIS_CLUSTER_NAME", "LOCAL")
-
-    System.setProperty("enhet.retry.max.retries", "3")
-    System.setProperty("enhet.retry.delay.seconds", "1")
-
-    val restClient = RestClient(
-        config = ClientConfig(scope = "AAP_SCOPES"),
-        tokenProvider = ClientCredentialsTokenProvider,
-        responseHandler = DefaultResponseHandler(),
-    )
-    motor.start()
-
-    val lagreAvklaringsbehovHendelseJobb = LagreAvklaringsbehovHendelseJobb(jobbAppender)
-
-    val server = embeddedServer(Netty, port = 0) {
-        module(
-            transactionExecutor,
-            jobbAppender,
-            azureConfig,
-            {},
-            lagreStoppetHendelseJobb,
-            lagreAvklaringsbehovHendelseJobb,
-        )
-    }.start()
-
-    val port = runBlocking { server.engine.resolvedConnectors().first().port }
-
-    res = TestClient(restClient, "http://localhost:$port").test()
-
-    motor.stop()
-    server.stop(500L, 10_000L)
-
-    return res
-}
-
-data class TestJobberSetup(
-    val lagreSakinfoTilBigQueryJobb: LagreSakinfoTilBigQueryJobb,
-    val lagreAvsluttetBehandlingTilBigQueryJobb: LagreAvsluttetBehandlingTilBigQueryJobb,
-    val resendSakstatistikkJobb: ResendSakstatistikkJobb,
-    val motorJobbAppender: MotorJobbAppender,
-)
-
-fun konstruerTestJobber(
-    bqYtelseRepository: IBQYtelsesstatistikkRepository = FakeBQYtelseRepository()
-): TestJobberSetup {
-    val lagreSakinfoTilBigQueryJobb = LagreSakinfoTilBigQueryJobb()
-    val lagreAvsluttetBehandlingTilBigQueryJobb =
-        LagreAvsluttetBehandlingTilBigQueryJobb(bqYtelseRepository)
-    val resendSakstatistikkJobb = ResendSakstatistikkJobb()
-    val motorJobbAppender = MotorJobbAppender()
-    return TestJobberSetup(
-        lagreSakinfoTilBigQueryJobb,
-        lagreAvsluttetBehandlingTilBigQueryJobb,
-        resendSakstatistikkJobb,
-        motorJobbAppender
-    )
-}
-
-fun konstruerMotor(
-    dataSource: DataSource,
-    jobbAppender: MotorJobbAppender,
-    bqYtelseRepository: IBQYtelsesstatistikkRepository,
-    resendSakstatistikkJobb: ResendSakstatistikkJobb,
-    lagreAvklaringsbehovHendelseJobb: LagreAvklaringsbehovHendelseJobb,
-    lagrePostmottakHendelseJobb: LagrePostmottakHendelseJobb,
-    lagreSakinfoTilBigQueryJobb: LagreSakinfoTilBigQueryJobb
-): Motor {
-    val lagreOppgaveJobb = LagreOppgaveJobb()
-    val lagreAvsluttetBehandlingTilBigQueryJobb =
-        LagreAvsluttetBehandlingTilBigQueryJobb(bqYtelseRepository)
-    return motor(
-        dataSource = dataSource,
-        gatewayProvider = defaultGatewayProvider { register<FakePdlGateway>() },
-        jobber = listOf(
-            lagreAvsluttetBehandlingTilBigQueryJobb,
-            lagreOppgaveJobb,
-            resendSakstatistikkJobb,
-            lagreAvklaringsbehovHendelseJobb,
-            lagrePostmottakHendelseJobb,
-            LagreOppgaveHendelseJobb(),
-            lagreSakinfoTilBigQueryJobb,
-            LagreStoppetHendelseJobb(jobbAppender, lagreAvsluttetBehandlingTilBigQueryJobb),
-            LagreTilbakekrevingHendelseJobb()
-        )
-    )
-}
-
-fun konstruerManuellMotor(
-    dataSource: DataSource,
-    jobbAppender: MotorJobbAppender,
-    bqYtelseRepository: IBQYtelsesstatistikkRepository,
-    resendSakstatistikkJobb: ResendSakstatistikkJobb,
-    lagreAvklaringsbehovHendelseJobb: LagreAvklaringsbehovHendelseJobb,
-    lagrePostmottakHendelseJobb: LagrePostmottakHendelseJobb,
-    lagreSakinfoTilBigQueryJobb: LagreSakinfoTilBigQueryJobb
-): ManuellMotorImpl {
-    val lagreOppgaveJobb = LagreOppgaveJobb()
-    val lagreAvsluttetBehandlingTilBigQueryJobb =
-        LagreAvsluttetBehandlingTilBigQueryJobb(bqYtelseRepository)
-    return ManuellMotorImpl(
-        dataSource = dataSource,
-        jobber = listOf(
-            lagreAvsluttetBehandlingTilBigQueryJobb,
-            lagreOppgaveJobb,
-            resendSakstatistikkJobb,
-            lagreAvklaringsbehovHendelseJobb,
-            lagrePostmottakHendelseJobb,
-            LagreOppgaveHendelseJobb(),
-            lagreSakinfoTilBigQueryJobb,
-            LagreStoppetHendelseJobb(jobbAppender, lagreAvsluttetBehandlingTilBigQueryJobb),
-            LagreTilbakekrevingHendelseJobb()
-        ),
-        repositoryRegistry = postgresRepositoryRegistry,
-        gatewayProvider = defaultGatewayProvider { register<FakePdlGateway>() },
-    )
-}
-
-@JvmOverloads
-fun opprettTestStoppetBehandling(
-    behandlingReferanse: UUID,
-    behandlingOpprettetTidspunkt: LocalDateTime,
-    hendelsesTidspunkt: LocalDateTime,
-    mottattTid: LocalDateTime,
-    saksnummer: String = "123",
-    behandlingStatus: Status = Status.OPPRETTET,
-    behandlingType: no.nav.aap.behandlingsflyt.kontrakt.behandling.TypeBehandling = no.nav.aap.behandlingsflyt.kontrakt.behandling.TypeBehandling.Førstegangsbehandling,
-    ident: String = "0",
-    avklaringsbehov: List<AvklaringsbehovHendelseDto> = listOf()
-): StoppetBehandling =
-    StoppetBehandling(
-        saksnummer = saksnummer,
-        behandlingStatus = behandlingStatus,
-        behandlingType = behandlingType,
-        ident = ident,
-        behandlingReferanse = behandlingReferanse,
-        behandlingOpprettetTidspunkt = behandlingOpprettetTidspunkt,
-        avklaringsbehov = avklaringsbehov,
-        versjon = "UKJENT",
-        mottattTid = mottattTid,
-        sakStatus = no.nav.aap.behandlingsflyt.kontrakt.sak.Status.UTREDES,
-        hendelsesTidspunkt = hendelsesTidspunkt,
-        vurderingsbehov = listOf(no.nav.aap.behandlingsflyt.kontrakt.statistikk.Vurderingsbehov.SØKNAD),
-        årsakTilOpprettelse = ÅrsakTilOpprettelse.SØKNAD
-    )
-
-object FakeBigQueryClient : IBigQueryClient {
-    override fun <E> create(table: BQTable<E>): Boolean {
-        return true
-    }
-
-    override fun <E> insert(table: BQTable<E>, value: E) = Unit
-
-    override fun <E> read(table: BQTable<E>): List<E> {
-        return emptyList()
-    }
-
-    override fun <E> read(
-        table: BQTable<E>,
-        whereClause: String
-    ): List<E> {
-        return emptyList()
-    }
-
-    override fun <E> insertMany(
-        table: BQTable<E>,
-        values: List<E>
-    ) = Unit
-}
-
-class TestClient(private val client: RestClient<InputStream>, private val url: String) {
-    fun postBehandlingsflytHendelse(
-        hendelse: StoppetBehandling
-    ) {
-        client.post<StoppetBehandling, Any>(
-            URI.create("$url/stoppetBehandling"), PostRequest(hendelse)
-        )
-    }
-
-    fun oppdatertBehandlingHendelse(hendelse: StoppetBehandling) {
-        client.post<StoppetBehandling, Any>(
-            URI.create("$url/oppdatertBehandling"), PostRequest(hendelse)
-        )
-    }
-
-    fun postOppgaveData(
-        oppgaveHendelse: OppgaveHendelse
-    ) {
-        client.post<OppgaveHendelse, Any>(
-            URI.create("$url/oppgave"), PostRequest(oppgaveHendelse)
-        )
-    }
-
-    fun postPostmottakHendelse(hendelse: DokumentflytStoppetHendelse) {
-        client.post<DokumentflytStoppetHendelse, Any>(
-            URI.create("$url/postmottak"), PostRequest(hendelse)
-        )
-    }
-
-    fun postTilbakekrevingshendelse(hendelse: TilbakekrevingsbehandlingOppdatertHendelse) {
-        client.post<TilbakekrevingsbehandlingOppdatertHendelse, Any>(
-            URI.create("$url/tilbakekrevingshendelse"), PostRequest(hendelse)
-        )
-    }
-}
-
-fun <E> testKlientNoInjection(
-    dbConfig: DbConfig,
-    azureConfig: AzureConfig = AzureConfig(
-        clientId = "tilgang",
-        jwksUri = "http://localhost:8081/jwks",
-        issuer = "tilgang"
-    ),
-    bigQueryClient: IBigQueryClient = FakeBigQueryClient,
-    test: TestClient.() -> E,
-): E {
-    var res: E
-
-    System.setProperty("azure.openid.config.token.endpoint", azureConfig.tokenEndpoint.toString())
-    System.setProperty("azure.app.client.id", azureConfig.clientId)
-    System.setProperty("azure.app.client.secret", azureConfig.clientSecret)
-    System.setProperty("azure.openid.config.jwks.uri", azureConfig.jwksUri)
-    System.setProperty("azure.openid.config.issuer", azureConfig.issuer)
-
-    System.setProperty("NAIS_CLUSTER_NAME", "LOCAL")
-
-    System.setProperty("enhet.retry.max.retries", "1")
-    System.setProperty("enhet.retry.delay.seconds", "1")
-
-
-    val restClient = RestClient(
-        config = ClientConfig(scope = "AAP_SCOPES"),
-        tokenProvider = ClientCredentialsTokenProvider,
-        responseHandler = DefaultResponseHandler()
-    )
-
-    val server = embeddedServer(Netty, port = 0) {
-        startUp(
-            dbConfig,
-            azureConfig,
-            bigQueryClient,
-            defaultGatewayProvider()
-        )
-    }.start()
-
-    val port = runBlocking { server.engine.resolvedConnectors().first().port }
-
-    res = TestClient(restClient, "http://localhost:$port").test()
-
-    server.stop(1000L, 10_000L)
-
-    return res
-}
-
-fun <E> testKlientNoInjectionManuell(
-    dbConfig: DbConfig,
-    azureConfig: AzureConfig = AzureConfig(
-        clientId = "tilgang",
-        jwksUri = "http://localhost:8081/jwks",
-        issuer = "tilgang"
-    ),
-    bigQueryClient: IBigQueryClient = FakeBigQueryClient,
-    test: TestClient.(ManuellMotorImpl) -> E,
-): E {
-    lateinit var motor: ManuellMotorImpl
-
-    System.setProperty("azure.openid.config.token.endpoint", azureConfig.tokenEndpoint.toString())
-    System.setProperty("azure.app.client.id", azureConfig.clientId)
-    System.setProperty("azure.app.client.secret", azureConfig.clientSecret)
-    System.setProperty("azure.openid.config.jwks.uri", azureConfig.jwksUri)
-    System.setProperty("azure.openid.config.issuer", azureConfig.issuer)
-
-    System.setProperty("NAIS_CLUSTER_NAME", "LOCAL")
-
-    System.setProperty("enhet.retry.max.retries", "1")
-    System.setProperty("enhet.retry.delay.seconds", "0")
-
-    val restClient = RestClient(
-        config = ClientConfig(scope = "AAP_SCOPES"),
-        tokenProvider = ClientCredentialsTokenProvider,
-        responseHandler = DefaultResponseHandler()
-    )
-
-    val server = embeddedServer(Netty, port = 0) {
-        startUp(
-            dbConfig,
-            azureConfig,
-            bigQueryClient,
-            defaultGatewayProvider()
-        ) { ds, gp, jobber ->
-            ManuellMotorImpl(ds, jobber, postgresRepositoryRegistry, gp).also { motor = it }
-        }
-    }.start()
-
-    val port = runBlocking { server.engine.resolvedConnectors().first().port }
-
-    val res = TestClient(restClient, "http://localhost:$port").test(motor)
-
-    server.stop(1000L, 10_000L)
-
-    return res
-}
-
-
-fun postgresTestConfig(): DbConfig {
-    val postgres = PostgreSQLContainer("postgres:16")
-    // Get the current working directory
-    val dumpFile = Path.of("").toAbsolutePath().resolve("dump.sql")
-
-    val forHostPath = MountableFile.forHostPath(dumpFile)
-
-    postgres.waitingFor(
-        HostPortWaitStrategy().withStartupTimeout(
-            Duration.of(
-                60L,
-                ChronoUnit.SECONDS
-            )
-        )
-    ).withCopyFileToContainer(
-        forHostPath,
-        "/dump.sql"
-    )
-    postgres.start()
-
-    if (dumpFile.toFile().exists()) {
-        val res = postgres.execInContainer("bash", "-c", "psql -U test -d test -f /dump.sql 2>&1")
-        if (res.exitCode != 0) {
-            println(res)
-        }
-    } else {
-        println("Kan ikke finne databasedump med path $dumpFile")
-    }
-
-    val dbConfig = DbConfig(
-        jdbcUrl = postgres.jdbcUrl,
-        userName = postgres.username,
-        password = postgres.password,
-        poolSize = AppConfig.hikariMaxPoolSize
-    )
-
-    return dbConfig
-}
-
-
-fun bigQueryContainer(): BigQueryConfig {
-    val container = BigQueryEmulatorContainer("ghcr.io/goccy/bigquery-emulator:0.6.3")
-    container.start()
-
-    val url = container.emulatorHttpEndpoint
-
-    val datasetInfo = DatasetInfo.newBuilder("my-dataset").build()
-    val config = object : BigQueryConfig {
-        override val dataset: String
-            get() = "my-dataset"
-
-        override fun bigQueryOptions(): BigQueryOptions {
-            return BigQueryOptions.newBuilder()
-                .setLocation(url)
-                .setProjectId(container.projectId)
-                .setHost(url)
-                .setCredentials(NoCredentials.getInstance())
-                .build()
-        }
-    }
-
-    logger.info("BigQuery URL: $url")
-
-    val bigQueryOptions: BigQueryOptions = config.bigQueryOptions()
-    val bigQuery: BigQuery = bigQueryOptions.service
-
-    // Lag datasett
-    bigQuery.create(datasetInfo)
-
-    return config
-}
-
-fun opprettTestHendelse(
-    dataSource: DataSource,
-    randomUUID: UUID,
-    saksnummer: Saksnummer,
-    status: BehandlingStatus = BehandlingStatus.UTREDES,
-    opprettetTidspunkt: LocalDateTime = LocalDateTime.now(),
-    vurderingsbehov: List<Vurderingsbehov> = emptyList(),
-    clock: Clock = Clock.systemDefaultZone()
-): Pair<BehandlingId, SakId> {
-    val ident = "29021946"
-
-    val personMedId = opprettTestPerson(dataSource, ident)
-
-    val sak = opprettTestSak(dataSource, saksnummer, Person(ident, id = personMedId.id()))
-
-    val behandling = opprettTestBehandling(
-        dataSource,
-        randomUUID,
-        sak,
-        status,
-        opprettetTidspunkt,
-        vurderingsbehov,
-        clock
-    )
-
-    val sakId = sak.id!!
-    val behandlingId = behandling.id()
-
-    return Pair(behandlingId, sakId)
-}
-
-fun opprettTestPerson(dataSource: DataSource, ident: String): Person {
-    return dataSource.transaction { conn ->
-        val personRepository = PersonRepository(conn)
-        PersonService(personRepository).hentEllerLagrePerson(ident)
-    }
-}
-
-fun opprettTestSak(dataSource: DataSource, saksnummer: Saksnummer, person: Person): Sak {
-    return dataSource.transaction {
-        val sak = Sak(
-            saksnummer = saksnummer,
-            person = person,
-            id = null,
-            sistOppdatert = LocalDateTime.now(),
-            sakStatus = SakStatus.UTREDES,
-        )
-        val id = SakRepositoryImpl(it).settInnSak(sak)
-        sak.copy(id = id)
-    }
-}
-
-fun opprettTestBehandling(
-    dataSource: DataSource,
-    referanse: UUID,
-    sak: Sak,
-    status: BehandlingStatus = BehandlingStatus.UTREDES,
-    opprettetTidspunkt: LocalDateTime = LocalDateTime.now(),
-    vurderingsbehov: List<Vurderingsbehov> = emptyList(),
-    clock: Clock = Clock.systemDefaultZone()
-): Behandling {
-    val behandling = Behandling(
-        referanse = referanse,
-        sak = sak,
-        typeBehandling = TypeBehandling.Førstegangsbehandling,
-        status = status,
-        opprettetTid = opprettetTidspunkt,
-        oppdatertTidspunkt = opprettetTidspunkt,
-        mottattTid = opprettetTidspunkt.truncatedTo(ChronoUnit.SECONDS),
-        versjon = Versjon(UUID.randomUUID().toString()),
-        årsaker = vurderingsbehov,
-        søknadsformat = SøknadsFormat.PAPIR,
-    )
-    return dataSource.transaction {
-        val repo = BehandlingRepository(it, clock = clock)
-        val uthentet = repo.hent(referanse)
-        val id = if (uthentet != null) {
-            repo.oppdaterBehandling(behandling.copy(id = uthentet.id))
-            uthentet.id
-        } else {
-            BehandlingRepository(it, clock).opprettBehandling(
-                behandling
-            )
-        }
-        behandling.copy(id = id)
-    }
-}
-
-val noOpTransactionExecutor = object : TransactionExecutor {
-    override fun <E> withinTransaction(block: (DBConnection) -> E): E {
-        return block(mockk(relaxed = true))
-    }
-}
-
-fun motorMock(): Motor {
-    val motor = mockk<Motor>()
-    every { motor.start() } just Runs
-    every { motor.stop() } just Runs
-    return motor
-}
-
-
-class MockJobbAppender : JobbAppender {
-    var jobber = mutableListOf<JobbInput>()
-    private var bigQueryJobber = mutableListOf<BehandlingId>()
-    var sisteEnhetRetryCount: Int = 0
-    var sisteDelayInSeconds: Long = 0
-    var sisteStoredBQBehandling: BQBehandling? = null
-    var sisteAvklaringsbehovKode: String? = null
-
-    override fun leggTil(
-        connection: DBConnection,
-        jobb: JobbInput
-    ) {
-        jobber.add(jobb)
-    }
-
-    override fun leggTil(
-        repositoryProvider: RepositoryProvider,
-        jobb: JobbInput
-    ) {
-        jobber.add(jobb)
-    }
-
-    override fun leggTilLagreSakTilBigQueryJobb(
-        repositoryProvider: RepositoryProvider,
-        behandlingId: BehandlingId,
-        delayInSeconds: Long,
-        enhetRetryCount: Int,
-        triggerKilde: String,
-        storedBQBehandling: BQBehandling?,
-        avklaringsbehovKode: String?,
-    ) {
-        logger.info("NO-OP: skal lagre til BigQuery for behandling $behandlingId. enhetRetryCount=$enhetRetryCount, delay=$delayInSeconds.")
-        bigQueryJobber.add(behandlingId)
-        sisteEnhetRetryCount = enhetRetryCount
-        sisteDelayInSeconds = delayInSeconds
-        sisteStoredBQBehandling = storedBQBehandling
-        sisteAvklaringsbehovKode = avklaringsbehovKode
-    }
-
-    override fun leggTilLagreAvsluttetBehandlingTilBigQueryJobb(
-        provider: RepositoryProvider,
-        behandlingId: BehandlingId,
-        lagreAvsluttetBehandlingTilBigQueryJobb: LagreAvsluttetBehandlingTilBigQueryJobb
-    ) {
-        TODO("Not yet implemented")
-    }
-
-    override fun leggTilResendSakstatistikkJobb(
-        repositoryProvider: RepositoryProvider,
-        behandlingId: BehandlingId
-    ) {
-        TODO("Not yet implemented")
-    }
-}
-
-class FakeMeldekortRepository : IMeldekortRepository {
-    private val meldekort = mutableMapOf<Long, List<Meldekort>>()
-    override fun lagre(
-        behandlingId: BehandlingId,
-        meldekort: List<Meldekort>
-    ) {
-        this.meldekort[behandlingId.id] = meldekort
-    }
-
-    override fun hentMeldekort(behandlingId: BehandlingId): List<Meldekort> {
-        return meldekort[behandlingId.id] ?: emptyList()
-    }
-}
-
-class FakeSakRepository : SakRepository {
-    private val saker = mutableMapOf<Long, Sak>()
-    override fun hentSak(sakID: SakId): Sak {
-        saker[sakID]?.let { return it }
-        throw IllegalArgumentException("Fant ikke sak med id $sakID")
-    }
-
-    override fun hentSak(saksnummer: Saksnummer): Sak {
-        return saker.values.firstOrNull { it.saksnummer == saksnummer }!!
-    }
-
-    override fun hentSakEllernull(saksnummer: Saksnummer): Sak? {
-        return saker.values.firstOrNull { it.saksnummer == saksnummer }
-            ?.also { requireNotNull(it.id) }
-    }
-
-    override fun settInnSak(sak: Sak): SakId {
-        val id = saker.size.toLong()
-        saker[id] = sak.copy(id = id)
-        return id
-    }
-
-    override fun oppdaterSak(sak: Sak) {
-        saker[sak.id!!] = sak.copy(sistOppdatert = LocalDateTime.now())
-    }
-
-    override fun tellSaker(): Int {
-        return saker.size
-    }
-}
-
-class FakePersonRepository : IPersonRepository {
-    private val personer = mutableMapOf<Long, Person>()
-    override fun lagrePerson(person: Person): Long {
-        personer[personer.size.toLong()] = person
-        return (personer.size - 1).toLong()
-    }
-
-    override fun hentPerson(ident: String): Person? {
-        return personer.values.firstOrNull { it.ident == ident }
-    }
-}
-
-class FakeBehandlingRepository : IBehandlingRepository {
-    private val behandlinger = mutableMapOf<Long, Behandling>()
-    private var nextId = 0L
-    override fun opprettBehandling(behandling: Behandling): BehandlingId {
-        val id = nextId
-        behandlinger[id] = behandling.copy(id = BehandlingId(id)).leggTilHendelse(
-            BehandlingHendelse(
-                tidspunkt = LocalDateTime.now(),
-                hendelsesTidspunkt = LocalDateTime.now(),
-                status = behandling.behandlingStatus(),
-                avklaringsbehovStatus = behandling.gjeldendeAvklaringsbehovStatus,
-                versjon = behandling.versjon,
-                mottattTid = behandling.mottattTid,
-                søknadsformat = behandling.søknadsformat,
-                relatertBehandlingReferanse = behandling.relatertBehandlingReferanse
-            )
-        )
-        nextId++
-
-        logger.info("Opprettet behandling med ID $id")
-        return BehandlingId(id)
-    }
-
-    override fun oppdaterBehandling(behandling: Behandling) {
-        logger.info("Oppdaterte behandling med ID ${behandling.id}")
-        behandlinger[behandling.id?.id!!] = behandling.leggTilHendelse(
-            BehandlingHendelse(
-                tidspunkt = LocalDateTime.now(),
-                hendelsesTidspunkt = LocalDateTime.now(),
-                status = behandling.behandlingStatus(),
-                avklaringsbehovStatus = behandling.gjeldendeAvklaringsbehovStatus,
-                versjon = behandling.versjon,
-                mottattTid = behandling.mottattTid,
-                søknadsformat = behandling.søknadsformat,
-                relatertBehandlingReferanse = behandling.relatertBehandlingReferanse
-            )
-        )
-    }
-
-    override fun invaliderOgLagreNyHistorikk(behandling: Behandling) {
-        behandlinger[behandling.id?.id!!] = behandling
-    }
-
-    override fun hent(referanse: UUID): Behandling? {
-        return behandlinger.asIterable().firstOrNull { it.value.referanse == referanse }?.value
-    }
-
-    override fun hent(id: BehandlingId): Behandling {
-        return behandlinger[id.id]!!
-    }
-
-    override fun hentEllerNull(id: BehandlingId): Behandling? {
-        return behandlinger[id.id]
-    }
-}
-
-class FakeRettighetsTypeRepository : IRettighetstypeperiodeRepository {
-    override fun lagre(
-        behandlingReferanse: UUID,
-        rettighetstypePeriode: List<RettighetstypePeriode>
-    ) {
-        TODO("Not yet implemented")
-    }
-
-    override fun hent(behandlingReferanse: UUID): List<RettighetstypePeriode> {
-        return emptyList()
-    }
-}
-
-class FakeArbeidsopptrappingRepository : ArbeidsopptrappingperioderRepository {
-    override fun lagre(
-        behandlingId: BehandlingId,
-        perioder: List<Periode>
-    ) {
-        TODO("Not yet implemented")
-    }
-
-    override fun hent(behandlingId: BehandlingId): List<Periode>? {
-        TODO("Not yet implemented")
-    }
-
-}
-
-class FakeVedtattStansOpphørRepository : VedtattStansOpphørRepository {
-    private val lagret = mutableMapOf<BehandlingId, List<StansEllerOpphør>>()
-
-    override fun lagre(behandlingId: BehandlingId, vedtattStansOpphør: List<StansEllerOpphør>) {
-        lagret[behandlingId] = vedtattStansOpphør
-    }
-
-    override fun hent(behandlingId: BehandlingId): List<StansEllerOpphør> =
-        lagret[behandlingId] ?: emptyList()
-}
-
-class FakeFritaksvurderingRepository : FritaksvurderingRepository {
-    override fun lagre(behandlingId: BehandlingId, vurderinger: List<Fritakvurdering>) {
-        TODO("Not yet implemented")
-    }
-
-    override fun hentFritaksvurderinger(behandlingId: BehandlingId): List<Fritakvurdering> {
-        TODO("Not yet implemented")
-    }
-}
-
-class FakeDiagnoseRepository : DiagnoseRepository {
-    override fun lagre(diagnoseEntity: DiagnoseEntity): Long {
-        TODO("Not yet implemented")
-    }
-
-    override fun hentForBehandling(behandlingReferanse: UUID): DiagnoseEntity {
-        TODO("Not yet implemented")
-    }
-}
-
-class FakeBQYtelseRepository : IBQYtelsesstatistikkRepository {
-    val behandlinger = mutableListOf<BQYtelseBehandling>()
-
-    override fun lagre(payload: BQYtelseBehandling) {
-        behandlinger.add(payload)
-    }
-
-    override fun commit() = Unit
-
-    override fun start() = Unit
-}
-
-class FakeTilkjentYtelseRepository : ITilkjentYtelseRepository {
-    private val tilkjentYtelser = mutableMapOf<Int, TilkjentYtelseEntity>()
-    override fun lagreTilkjentYtelse(tilkjentYtelse: TilkjentYtelseEntity): Long {
-        tilkjentYtelser[tilkjentYtelser.size] = tilkjentYtelse
-        return (tilkjentYtelser.size - 1).toLong()
-    }
-
-    override fun hentTilkjentYtelse(tilkjentYtelseId: Int): TilkjentYtelse {
-        TODO("Not yet implemented")
-    }
-
-    override fun hentForBehandling(behandlingId: UUID): TilkjentYtelse {
-        TODO("Not yet implemented")
-    }
-}
-
-class FakeVilkårsResultatRepository : IVilkårsresultatRepository {
-    private val vilkår = mutableMapOf<Long, VilkårsResultatEntity>()
-
-    override fun lagreVilkårsResultat(
-        vilkårsresultat: VilkårsResultatEntity,
-        behandlingId: BehandlingId
-    ): Long {
-        vilkår[vilkår.size.toLong()] = vilkårsresultat
-        return (vilkår.size - 1).toLong()
-    }
-
-    override fun hentVilkårsResultat(vilkårResultatId: Long): VilkårsResultatEntity? {
-        TODO("Not yet implemented")
-    }
-
-    override fun hentForBehandling(behandlingsReferanse: UUID): VilkårsResultatEntity {
-        TODO("Not yet implemented")
-    }
-}
-
-class FakeBeregningsgrunnlagRepository : IBeregningsgrunnlagRepository {
-    val grunnlag = mutableListOf<MedBehandlingsreferanse<IBeregningsGrunnlag>>()
-    override fun lagreBeregningsGrunnlag(beregningsGrunnlag: MedBehandlingsreferanse<IBeregningsGrunnlag>): Long {
-        grunnlag.add(beregningsGrunnlag)
-        return grunnlag.indexOf(beregningsGrunnlag).toLong()
-    }
-
-    override fun hentBeregningsGrunnlag(referanse: UUID): List<MedBehandlingsreferanse<IBeregningsGrunnlag>> {
-        return grunnlag
-    }
-}
-
-class FakePdlGateway(val identerHemmelig: Map<String, Boolean> = emptyMap()) : PdlGateway {
-    companion object : Factory<PdlGateway> {
-        override fun konstruer(): PdlGateway {
-            return FakePdlGateway()
-        }
-    }
-
-    override fun hentPersoner(identer: List<String>): List<no.nav.aap.statistikk.integrasjoner.pdl.Person> {
-        return identer.map {
-            no.nav.aap.statistikk.integrasjoner.pdl.Person(
-                adressebeskyttelse = listOf(Adressebeskyttelse(gradering = if (identerHemmelig[it] == true) Gradering.STRENGT_FORTROLIG else Gradering.UGRADERT))
-            )
-        }
-    }
-}
-
-class FakeOppgaveHendelseRepository : OppgaveHendelseRepository {
-    private val enhetReservasjoner =
-        mutableMapOf<Pair<UUID, String>, List<EnhetReservasjonOgTidspunkt>>()
-
-    fun addEnhetReservasjon(
-        behandlingRef: UUID,
-        avklaringsbehovKode: String,
-        data: List<EnhetReservasjonOgTidspunkt>
-    ) {
-        enhetReservasjoner[behandlingRef to avklaringsbehovKode] = data
-    }
-
-    override fun hentEnhetOgReservasjonForAvklaringsbehov(
-        behandlingReferanse: UUID,
-        avklaringsbehovKode: String
-    ): List<EnhetReservasjonOgTidspunkt> {
-        return enhetReservasjoner[behandlingReferanse to avklaringsbehovKode] ?: emptyList()
-    }
-
-    override fun lagreHendelse(hendelse: no.nav.aap.statistikk.oppgave.OppgaveHendelse) = 0L
-    override fun sisteVersjonForId(id: Long) = null
-    override fun hentHendelserForId(id: Long) =
-        emptyList<no.nav.aap.statistikk.oppgave.OppgaveHendelse>()
-}
-
-fun <E> ventPåSvar(getter: () -> E?, predicate: (E?) -> Boolean): E? {
-    var res: E? = null
-    val timeInMillis = measureTimeMillis {
-        val maxTid = LocalDateTime.now().plusSeconds(10)
-        var suksess = false
-        while (maxTid.isAfter(LocalDateTime.now()) && !suksess) {
-            try {
-                res = getter()
-                if (res != null && predicate(res)) {
-                    suksess = true
-                }
-            } finally {
-                Thread.sleep(50L)
-            }
-        }
-    }
-    logger.info("Ventet på at prosessering skulle fullføre, det tok $timeInMillis millisekunder. Res null: ${res == null}")
-    if (res == null) {
-        logger.info("Ventet på svar, men svaret er null.")
-    }
-    return res
-}
-
-/**
- * Venter til alle jobber er ferdige, men avbryter umiddelbart hvis noen jobber feiler.
- * Forhindrer lang ventetid i integrasjonstester ved uventede jobbkrasj.
- */
-fun ventPåSvarEllerFeil(
-    dataSource: DataSource,
-    cronJobberSomSkalIgnoreres: List<String> = listOf("oppgave.retryFeilede"),
-    maxTidSekunder: Long = 20,
-) {
-    val sluttTidspunkt = LocalDateTime.now().plusSeconds(maxTidSekunder)
-    while (LocalDateTime.now().isBefore(sluttTidspunkt)) {
-        val (feilende, harVentende) = dataSource.transaction(readOnly = true) { connection ->
-            val feilende = DriftJobbRepositoryExposed(connection).hentAlleFeilende()
-            val harVentende = TestJobbRepository(connection, cronJobberSomSkalIgnoreres).harJobb(null, null)
-            Pair(feilende, harVentende)
-        }
-
-        if (feilende.isNotEmpty()) {
-            val detaljer = feilende.joinToString("\n") { (jobb, melding) -> "  ${jobb.type()}: $melding" }
-            throw AssertionError("${feilende.size} jobber feilet:\n$detaljer")
-        }
-
-        if (!harVentende) return
-
-        Thread.sleep(50)
-    }
-    throw AssertionError("Timeout: jobber ikke ferdig etter $maxTidSekunder sekunder")
-}
-
-fun forberedDatabase(
-    it: DBConnection,
-    behandlingReferanse: UUID
-): BehandlingId {
-    val ident = "214"
-    val person = PersonService(PersonRepository(it)).hentEllerLagrePerson(ident)
-
-    val sak = Sak(
-        saksnummer = "ABCDE".tilSaksnummer(),
-        person = person,
-        sakStatus = SakStatus.LØPENDE,
-        sistOppdatert = LocalDateTime.now()
-    )
-    val sakId = SakRepositoryImpl(it).settInnSak(sak)
-
-    return BehandlingRepository(it).opprettBehandling(
-        Behandling(
-            referanse = behandlingReferanse,
-            sak = sak.copy(id = sakId),
-            typeBehandling = TypeBehandling.Førstegangsbehandling,
-            status = BehandlingStatus.OPPRETTET,
-            opprettetTid = LocalDateTime.now(),
-            mottattTid = LocalDateTime.now().minusDays(1).truncatedTo(ChronoUnit.SECONDS),
-            versjon = Versjon("xxx"),
-            søknadsformat = SøknadsFormat.DIGITAL,
-            oppdatertTidspunkt = LocalDateTime.now()
-        )
-    )
-}
-
-
-fun konstruerSakstatistikkService(
-    connection: DBConnection
-): SaksStatistikkService {
-    val behandlingService = BehandlingService(
-        BehandlingRepository(connection),
-        SkjermingService(FakePdlGateway())
-    )
-    return SaksStatistikkService(
-        behandlingService = behandlingService,
-        sakstatistikkRepository = SakstatistikkRepositoryImpl(connection),
-        bqBehandlingMapper = BQBehandlingMapper(
-            behandlingService,
-            RettighetstypeperiodeRepository(connection),
-            OppgaveRepositoryImpl(connection),
-            SakstatistikkEventSourcing()
-        )
-    )
-}
-
-
-val schemaRegistry: SchemaRegistry = schemaRegistryYtelseStatistikk
